+++ source/renderer/app/containers/wallet/WalletSummaryPage.js
@@ -1,13 +1,13 @@
 // @flow
 import React, { Component } from 'react';
-import { get } from 'lodash';
+import { get, take } from 'lodash';
 import { observer, inject } from 'mobx-react';
 import { defineMessages, intlShape } from 'react-intl';
+import { MAX_TRANSACTIONS_ON_SUMMARY_PAGE } from '../../config/numbersConfig';
 import WalletTransactionsList from '../../components/wallet/transactions/WalletTransactionsList';
 import WalletSummary from '../../components/wallet/summary/WalletSummary';
 import WalletNoTransactions from '../../components/wallet/transactions/WalletNoTransactions';
 import VerticalFlexContainer from '../../components/layout/VerticalFlexContainer';
-import { DECIMAL_PLACES_IN_ADA } from '../../config/numbersConfig';
 import { ROUTES } from '../../routes-config';
 import type { InjectedProps } from '../../types/injectedPropsType';
 import { formattedWalletAmount } from '../../utils/formatters';
@@ -30,6 +30,13 @@ export default class WalletSummaryPage extends Component<Props> {
 
   static contextTypes = {
     intl: intlShape.isRequired,
+  };
+
+  handleShowMoreTransaction = (walletId: string) => {
+    this.props.actions.router.goToRoute.trigger({
+      route: ROUTES.WALLETS.PAGE,
+      params: { id: walletId, page: 'transactions' },
+    });
   };
 
   render() {
@@ -56,17 +63,18 @@ export default class WalletSummaryPage extends Component<Props> {
       walletTransactions = (
         <WalletTransactionsList
           key={`WalletTransactionsList_${wallet.id}`}
-          transactions={recent}
+          transactions={take(recent, MAX_TRANSACTIONS_ON_SUMMARY_PAGE)}
           isLoadingTransactions={recentTransactionsRequest.isExecutingFirstTime}
           hasMoreToLoad={false}
           assuranceMode={wallet.assuranceMode}
           walletId={wallet.id}
           isRestoreActive={isRestoreActive}
           formattedWalletAmount={formattedWalletAmount}
-          showMoreTransactionsButton={totalAvailable > 5}
+          showMoreTransactionsButton={recent.length > MAX_TRANSACTIONS_ON_SUMMARY_PAGE}
           network={network}
           onOpenExternalLink={openExternalLink}
           onShowMoreTransactions={this.handleShowMoreTransaction}
+          totalAvailable={totalAvailable}
         />
       );
     } else if (!hasAny) {
@@ -76,9 +84,8 @@ export default class WalletSummaryPage extends Component<Props> {
     return (
       <VerticalFlexContainer>
         <WalletSummary
-          walletName={wallet.name}
-          amount={wallet.amount.toFormat(DECIMAL_PLACES_IN_ADA)}
-          numberOfTransactions={totalAvailable}
+          wallet={wallet}
+          numberOfTransactions={totalAvailable || recent.length}
           pendingAmount={unconfirmedAmount}
           isLoadingTransactions={recentTransactionsRequest.isExecutingFirstTime}
           isRestoreActive={isRestoreActive}
@@ -87,11 +94,4 @@ export default class WalletSummaryPage extends Component<Props> {
       </VerticalFlexContainer>
     );
   }
-
-  handleShowMoreTransaction = (walletId: string) => {
-    this.props.actions.router.goToRoute.trigger({
-      route: ROUTES.WALLETS.PAGE,
-      params: { id: walletId, page: 'transactions' },
-    });
-  };
 }